switch to pg Pool, finish toggleWatchList, buystock handlers

+++ server/postgres/tables/transactions.sql
@@ -2,7 +2,7 @@ BEGIN TRANSACTION;
 CREATE TABLE transactions (
   transaction_id TEXT primary key,
   symbol VARCHAR(10),
-  price INT,
+  price FLOAT,
   quantity INT
 );
 COMMIT;
